ai-solution-finder-api/requirements.txt
fastapi
uvicorn
openai
langchain-openai
langchain
faiss-cpu
pypdf
python-dotenv
python-multipart
supabase